Add a compile-time option to use array indexes instead of pointers as node addresses.
The theory is that a 4-byte integer is smaller than an 8-byte pointer, and so nodes will be
smaller and more of them will fit in L1 cache.  In practice, it doesn't seem to make much
difference since most test problems fit in L1 cache already.  Nevertheless, having it as an
option lets us develop in both modes and possibly make the code more robust as a result.